requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
xarray>=2022.3.0
netCDF4>=1.6.0
h5py>=3.7.0
requests>=2.28.0
beautifulsoup4>=4.11.0
folium>=0.14.0
plotly>=5.15.0
cartopy>=0.21.0
pyproj>=3.4.0
rasterio>=1.3.0
geopandas>=0.12.0
tqdm>=4.64.0
python-dotenv>=1.0.0